add fastball executable
Using rake tasks is ok for development, but not reliable enough for
deployment. For example, running a fastball rake task in a rails app in
which you have not yet created config/database.yml will result in an
error and your config files will never get generated. This is what
happens when you use a deployment tool, like Ansible, to run the
fastball rake tasks during a deploy. Instead, you can now use the
fastball executable to generate config files with your deploy tool of
choice.